Explicitly enable the New Architecture
The New Architecture comes with the Interop Layer which allows
legacy native module library to work without changing any code.

According to https://github.com/reactwg/react-native-new-architecture/blob/main/docs/enable-libraries.md
the first step is to support the New Architecture is to make sure the
library works with the Interop Layer.

I tested that the legacy native module works fine with the Interop Layer.
In particular, sending events from native to JavaScript work as well.

So our SDK is compatible with the Interop Layer.
It is ready for adding support for the New Architecture.